+++ autolens/interferometer/model/aggregator.py
@@ -4,7 +4,7 @@ import autofit as af
 import autoarray as aa
 import autogalaxy as ag
 
-from autogalaxy.analysis.aggregator.aggregator import _interferometer_from
+from autogalaxy.interferometer.model.aggregator import _interferometer_from
 
 from autolens.interferometer.fit_interferometer import FitInterferometer
 from autolens.lens.model.aggregator.aggregator import AbstractAgg
