+++ autolens/lens/model/aggregator/__init__.py
@@ -1,8 +1,8 @@
-from autogalaxy.analysis.aggregator.aggregator import _imaging_from
-from autogalaxy.analysis.aggregator.aggregator import ImagingAgg
+from autogalaxy.imaging.model.aggregator import _imaging_from
+from autogalaxy.imaging.model.aggregator import ImagingAgg
 
-from autogalaxy.analysis.aggregator.aggregator import _interferometer_from
-from autogalaxy.analysis.aggregator.aggregator import InterferometerAgg
+from autogalaxy.interferometer.model.aggregator import _interferometer_from
+from autogalaxy.interferometer.model.aggregator import InterferometerAgg
 
 from autolens.lens.model.aggregator.aggregator import _tracer_from
 from autolens.lens.model.aggregator.aggregator import TracerAgg
@@ -10,7 +10,7 @@ from autolens.lens.model.aggregator.aggregator import TracerAgg
 from autolens.imaging.model.aggregator import _fit_imaging_from
 from autolens.imaging.model.aggregator import FitImagingAgg
 
-from autolens.lens.model.aggregator.aggregator import _interferometer_from
-from autolens.lens.model.aggregator.aggregator import FitInterferometerAgg
+from autolens.interferometer.model.aggregator import _fit_interferometer_from
+from autolens.interferometer.model.aggregator import FitInterferometerAgg
 
 from autolens.lens.model.aggregator.aggregator import SubhaloAgg
