+++ autolens/lens/plot/ray_tracing_plotters.py
