+++ test_autolens/lens/plot/test_ray_tracing_plotters.py
